+++ src/main/resources/lua/update_cart_menu.lua
@@ -1,6 +1,6 @@
 local currentQuantityStr = redis.call('HGET', KEYS[1], KEYS[2])
 local currentQuantity = tonumber(currentQuantityStr) or 0
-local newQuantity = tonumber(ARGV[1]) or 0
+local newQuantity = tonumber(ARGV[1])
 local unitPrice = tonumber(ARGV[2])
 
 -- 음수 검증
@@ -14,13 +14,14 @@ if currentQuantity == 0 then
 end
 
 -- 메뉴 개수 수정
+local changePrice = 0
+
 if newQuantity == 0 then
 	redis.call('HDEL', KEYS[1], KEYS[2])
 else
 	redis.call('HSET', KEYS[1], KEYS[2], newQuantity)
+	changePrice = (newQuantity - currentQuantity) * unitPrice
 end
-
-local changePrice = (newQuantity - currentQuantity) * unitPrice
 
 -- 총 금액 계산
 redis.call('HINCRBY', KEYS[1], KEYS[3], changePrice)
